pyproject.toml
[project]
name = "transcriber"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "ffmpeg-python>=0.2.0",
    "gradio>=5.49.1",
    "huggingface>=0.0.1",
    "ipython>=9.6.0",
    "modal>=1.2.1",
    "python-dotenv>=1.1.1",
    "pytubefix>=10.0.1",
    "transformers>=4.57.1",
]
